Move the common semantic layer node components to v1 artifact resources
