+++ core/dbt/artifacts/resources/v1/semantic_layer_components.py
